+++ src/pymetropolis/metro_network/road_network/files.py
@@ -273,6 +273,12 @@ class RoadEdgesPenaltiesFile(MetroDataFrameFile):
             description="Constant time penalty of the edge, in seconds.",
             nullable=True,
         ),
+        Column(
+            "speed_multiplier",
+            MetroDataType.FLOAT,
+            description="By how much edge speed limit is multiplied to get edge free-flow speed.",
+            nullable=True,
+        ),
     ]